requirements.txt
boto3>=1.21.42
python-dotenv>=0.20.0
tqdm>=4.64.0